+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   `Migrate` to the `FormsClient` class
+
 ## 2.7.0 (2021-12-21)
 
 ### Added